src/test/java/com/grasia/prima/ppi/api/helper/entity/GalleryHelperTest.java
package com.grasia.prima.ppi.api.helper.entity;

import com.grasia.prima.ppi.api.constant.GlobalMessage;
import com.grasia.prima.ppi.api.dto.response.GalleryResponse;
import com.grasia.prima.ppi.api.exception.BusinessException;
import org.junit.jupiter.api.Test;

import java.lang.reflect.Constructor;
import java.lang.reflect.InvocationTargetException;

import static org.junit.jupiter.api.Assertions.*;

class GalleryHelperTest {

    @Test
    void testInstanceGalleryHelper() throws NoSuchMethodException {
        Constructor<GalleryHelper> constructor = GalleryHelper.class.getDeclaredConstructor();
        constructor.setAccessible(true);
        InvocationTargetException e = assertThrows(InvocationTargetException.class, constructor::newInstance);
        assertTrue(e.getCause() instanceof BusinessException);
        assertEquals(GlobalMessage.INTERNAL_SERVER_ERROR.message, e.getCause().getMessage());
    }

    @Test
    void testToGalleryResponse_GalleryIsNull() {
        GalleryResponse response = GalleryHelper.toGalleryResponse(null);
        assertNull(response);
    }
}